move TraceDiGraph into separate file
